Move resource intesive tests in test_ham_density.
